Logic added to start a new round when a word has been finished
The words are shown always in the same order. The balloon screen is
cleared when a word has been completely written and a new round with
a new word is started.

The words have different length and the proper updating of the
word panel is not yet working. The rectangles surrounding each
word overlap if a longer word follows a shorter word.